Refactor NewItems card into component file
